proposed changes for issue #619
