+++ tljh/jupyterhub_config.py
@@ -6,42 +6,11 @@ from glob import glob
 import os
 
 from systemdspawner import SystemdSpawner
-from tljh import configurer, user
+from tljh import configurer
 from tljh.config import INSTALL_PREFIX, USER_ENV_PREFIX, CONFIG_DIR
-from tljh.normalize import generate_system_username
 from tljh.utils import get_plugin_manager
+from tljh.user_creating_spawner import UserCreatingSpawner
 from jupyterhub_traefik_proxy import TraefikTomlProxy
-
-from traitlets import Dict, Unicode, List
-
-class UserCreatingSpawner(SystemdSpawner):
-    """
-    SystemdSpawner with user creation on spawn.
-
-    FIXME: Remove this somehow?
-    """
-    user_groups = Dict(key_trait=Unicode(), value_trait=List(Unicode()), config=True)
-
-    def start(self):
-        """
-        Perform system user activities before starting server
-        """
-        # FIXME: Move this elsewhere? Into the Authenticator?
-        system_username = generate_system_username('jupyter-' + self.user.name)
-
-        # FIXME: This is a hack. Allow setting username directly instead
-        self.username_template = system_username
-        user.ensure_user(system_username)
-        user.ensure_user_group(system_username, 'jupyterhub-users')
-        if self.user.admin:
-            user.ensure_user_group(system_username, 'jupyterhub-admins')
-        else:
-            user.remove_user_group(system_username, 'jupyterhub-admins')
-        if self.user_groups:
-            for group, users in self.user_groups.items():
-                if self.user.name in users:
-                    user.ensure_user_group(system_username, group)
-        return super().start()
 
 c.JupyterHub.spawner_class = UserCreatingSpawner
